fix(estoque): corrige bug entrada com decimais e adiciona migracao sqlite automatica; refatora layout entrada de insumo

+++ backend/app/models/entrada_insumo.py
@@ -6,7 +6,7 @@ class EntradaInsumo(SQLModel, table=True):
     id: Optional[int] = Field(default=None, primary_key=True)
     insumo_id: int = Field(foreign_key="insumo.id")
     valor_aquisicao: float
-    quantidade: int
+    quantidade: float
     data_aquisicao: datetime = Field(default_factory=datetime.utcnow)
     condicao_pagamento: Optional[str] = Field(default="À Vista")
     data_pagamento: Optional[datetime] = None
